+++ src/tooling/piral-cli/src/common/port.ts
@@ -1,11 +1,17 @@
-import { log } from './log';
+import { log, fail } from './log';
 import { getPort } from '../external';
 
-export async function getAvailablePort(defaultPort: number) {
+export async function getAvailablePort(defaultPort: number, strict: boolean) {
   const selectedPort = await getFreePort(defaultPort);
 
   if (selectedPort !== defaultPort) {
-    log('portNotFree_0047', selectedPort, defaultPort);
+    if (strict) {
+      // exit
+      fail('portNotFree_0048', defaultPort);
+    } else {
+      // just print warning
+      log('portChanged_0047', selectedPort, defaultPort);
+    }
   }
 
   return selectedPort;